Remove an ugly cast and fix a compiler warning
compat_getopt_main.c: In function ‘handle’:
compat_getopt_main.c:123: warning: format ‘%s’ expects type ‘char *’, but argument 3 has type ‘long unsigned int’


git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@4742 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/common/compat_getopt_main.c
@@ -119,8 +119,7 @@ handle(char *progname,
     {
       fprintf(stderr,
               "%s: %lu bytes copied from `%s' to `%s'\n",
-              (unsigned long) progname, bytes_copied, infilename,
-              outfilename);
+              progname, bytes_copied, infilename, outfilename);
     }
   return 0;
 }
